requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.7.0
ta==0.10.2
schedule==1.2.1
yfinance==0.2.36
apscheduler==3.10.4
mplfinance==0.12.10b0 